// src/entities/ITEMS/TV-item/types/TV-item.types.ts
export type TVSubNames = 
    | "TV_Haier"
    | "TV_LG"
    | "TV_Panasonic"
    | "TV_Philips"
    | "TV_Samsung"
    | "TV_Sony"
    | "TV_Toshiba"

export type TVTypes = {
    id: number
    price: number
    name: TVNames
    subName: TVSubNames
    equipment: TVEquipment
    defects: TVDefects
}

export type TVEquipment = {
    powerCord: number
    remoteController: number
    originalBox: number
}

export type TVDefects = {
    scratchesBody: number
    scratchesFrame: number
    brokenButtons: number
    crackedChipScreen: number
    coloredWhiteDotsScreen: number
    scratchesDisplay: number
    doesntConnectWifi: number
    notCleatSound: number
    dontWorkSpeakers: number
    turnsOffItself: number
}